refactor: modified directory structure
