+++ frontend/src/app/components/images/GalleryImage.tsx
