+++ frontend/src/app/util/styleConstants.ts
